pyproject.toml
[tool.poetry]
name = "skillhub-api"
version = "0.1.0"
description = ""
authors = ["khoshov"]

[tool.poetry.dependencies]
python = "^3.9"
django = "4.2.4"
django-import-export = "3.2.0"
django-mptt = "^0.14.0"
django-ninja = "0.22.2"
pytils = "0.4.1"
django-ckeditor = "^6.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
